finalisation du projet

+++ src/main/java/com/application/safetynet/controller/FireStationController.java
@@ -1,6 +1,7 @@
 package com.application.safetynet.controller;
 
 import com.application.safetynet.exceptions.AddressNotExistException;
+import com.application.safetynet.exceptions.BadRequestException;
 import com.application.safetynet.exceptions.FireStationNotExistException;
 import com.application.safetynet.model.FireStation;
 import com.application.safetynet.model.dto.CountChildAndAdultDto;
@@ -12,7 +13,6 @@ import com.application.safetynet.repository.PersonRepository;
 import com.application.safetynet.service.FireStationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -25,14 +25,18 @@ import java.util.Optional;
 public class FireStationController {
     private static final Logger logger = LoggerFactory.getLogger(FireStationController.class);
 
-    @Autowired
     private FireStationService fireStationService;
 
-    @Autowired
     private FireStationRepository fireStationRepository;
 
-    @Autowired
     private PersonRepository personRepository;
+
+
+    public FireStationController(FireStationService fireStationService, FireStationRepository fireStationRepository, PersonRepository personRepository) {
+        this.fireStationService = fireStationService;
+        this.fireStationRepository = fireStationRepository;
+        this.personRepository = personRepository;
+    }
 
     /**
      * Get all persons attached to the station with a count of people under 18 and under 18
@@ -81,11 +85,15 @@ public class FireStationController {
         return fireStationService.getFamilyByStation(id);
     }
 
+    @GetMapping("/firestation/findall")
+    public List<FireStation> findAll () throws IOException {
+        return fireStationService.findAll();
+    }
+
     @PostMapping("/firestation")
     public FireStation addFireStation(@Valid @RequestBody FireStationDto fireStationDto) {
         logger.debug("Creating firestation {}", fireStationDto);
         List<String> address = new ArrayList<>();
-
         address.add(fireStationDto.getAddress());
         FireStation convertFireStationDto = new FireStation(fireStationDto.getStation(), address);
         return fireStationService.saveFireStation(convertFireStationDto);
@@ -118,7 +126,18 @@ public class FireStationController {
 
     @DeleteMapping("/firestation")
     public void deleteFireStation(@RequestBody FireStationDto id) {
-        logger.debug("Deleting firestation {}", id);
-        fireStationService.deleteFireStations(id);
+        if(id.getAddress() != null && id.getStation() != null) throw new BadRequestException("Veuillez renseigner soit l'addresse, soit le numero de station mais pas les deux");
+        else if(id.getAddress() != null){
+            boolean addressExist = fireStationRepository.findAll().stream().anyMatch(element -> element.getAddresses().contains(id.getAddress()));
+            if (!addressExist) throw new AddressNotExistException("L'adresse "+id.getAddress()+" que vous avez saisie n'existe pas dans la base de donnée");
+            logger.debug("Deleting firestation {}", id);
+            fireStationService.deleteFireStations(id);
+        }
+        else if(id.getStation() != null){
+            boolean stationExist = fireStationRepository.findAll().stream().anyMatch(station -> station.getStation().equalsIgnoreCase(id.getStation()));
+            if (!stationExist) throw new FireStationNotExistException("La station " + id.getStation() + " que vous avez saisie n'existe pas");
+            logger.debug("Deleting firestation {}", id);
+            fireStationService.deleteFireStations(id);
+        }
     }
 }
